Add enchantment loading and entity AI behavior systems
- Implemented loading of enchantments from JSON into NBT format in `valence_registry`.
- Created a modular entity AI system in `valence_vanilla` with behavior trees, including sequence, selector, and condition nodes.
- Added redstone component processing in `valence_vanilla`, handling various redstone blocks like wires, torches, repeaters, comparators, lamps, and pistons.
- Introduced a new `extract_jar` utility to extract specific JSON data from the server JAR file for registry entries.

+++ crates/valence_vanilla/src/combat.rs
@@ -1,11 +1,7 @@
-use std::time::Duration;
-
 use bevy_app::prelude::*;
 use bevy_ecs::prelude::*;
 use valence_entity::Position;
 use valence_math::DVec3;
-use valence_protocol::BlockPos;
-use valence_server::Client;
 
 pub struct CombatPlugin;
 
@@ -303,9 +299,9 @@ pub fn calculate_knockback(
         let nz = diff.z / horizontal_dist;
 
         DVec3::new(
-            -nx * factor + (rand::random::<f64>() - 0.5) * 0.2,
+            -nx * factor as f64 + (rand::random::<f64>() - 0.5) * 0.2,
             0.3 * factor as f64,
-            -nz * factor + (rand::random::<f64>() - 0.5) * 0.2,
+            -nz * factor as f64 + (rand::random::<f64>() - 0.5) * 0.2,
         )
     } else {
         DVec3::new(0.0, 0.3, 0.0)
